requirements.txt
actionlib==1.14.0
albucore==0.0.17
albumentations==1.4.18
angles==1.9.13
annotated-types==0.7.0
attrs==25.1.0
base-local-planner==1.17.3
blinker==1.8.2
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
Cython==3.0.12
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
dynamixel-sdk==3.7.51
easydict==1.13
easyocr==1.7.2
eval-type-backport==0.2.2
exceptiongroup==1.2.2
filelock==3.16.1
flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
h11==0.14.0
humanfriendly==10.0
idna==3.10
image-geometry==1.16.2
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
insightface==0.7.3
interactive-markers==1.12.0
itsdangerous==2.2.0
jinja2==3.1.5
joblib==1.4.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.7
laser-geometry==1.6.7
lazy-loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
message-filters==1.16.0
moveit-core==1.1.13
moveit-ros-planning-interface==1.1.13
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxruntime==1.16.3
opencv-python-headless==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
prettytable==3.11.0
protobuf==5.29.3
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic-core==2.27.2
pyparsing==3.1.4
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytz==2025.1
PyWavelets==1.4.1
PyYAML==6.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.32.3
resource-retriever==1.12.7
ros-numpy==0.0.5
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rviz==1.14.20
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.27.1
sensor-msgs==1.13.1
shapely==2.0.7
six==1.17.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sniffio==1.3.1
sortedcontainers==2.4.0
sound-play==0.3.17
srdfdom==0.6.4
sympy==1.13.3
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tf2-sensor-msgs==0.7.7
threadpoolctl==3.5.0
tifffile==2023.7.10
topic-tools==1.16.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
triton==3.0.0
typing-extensions==4.12.2
tzdata==2025.1
urdfdom-py==0.4.6
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
werkzeug==3.0.6
wsproto==1.2.0
xacro==1.14.16
zipp==3.20.2
